+++ spec/broadcaster/acts_as_publisher_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
 
 RSpec.describe Broadcaster::ActsAsPublisher do
-  describe '#publication_campaigns' do
+  describe '#broadcaster_campaigns' do
     let(:publisher) { create :user }
 
-    it { expect(publisher).to respond_to(:publication_campaigns) }
+    it { expect(publisher).to respond_to(:broadcaster_campaigns) }
   end
 end