+++ spec/broadcaster/acts_as_subscriber_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
 
 RSpec.describe Broadcaster::ActsAsSubscriber do
-  describe '#subscriptions' do
+  describe '#broadcaster_subscriptions' do
     let(:subscriber) { create :user }
 
-    it { expect(subscriber).to respond_to(:subscriptions) }
+    it { expect(subscriber).to respond_to(:broadcaster_subscriptions) }
   end
 end