+++ spec/dummy/app/models/user.rb
@@ -1,5 +1,5 @@
 # A model class for the sample application.
 class User < ActiveRecord::Base
-  acts_as_subscriber
-  acts_as_publisher
+  acts_as_broadcaster_subscriber
+  acts_as_broadcaster_publisher
 end