+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This project adheres to [Semantic Versioning](http://semver.org)
 
+## [0.2.0] - 2019-06-18
+
+Breaking changes:
+ - acts_as_subscriber => acts_as_broadcaster_subscriber
+ - acts_as_publisher => acts_as_broadcaster_publisher
+ - `publication_campaigns` => `broadcaster_campaign`
+ - `subscriptions` => `broadcaster_subscriptions`
+
 ## [0.1.2] - 2019-06-18
 
 Bugfixes: